requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.5.1
pydantic_core==2.14.3
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
